add putPersons to Dialog, fix Response/request Apis

+++ src/main/java/ru/skillbox/socialnetwork/api/request/DialogUsersApi.java
@@ -5,13 +5,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class DialogUsersApi {
 
    @JsonProperty("user_ids")
-   private long[] userIds;
+   private int[] userIds;
 
-   public long[] getUserIds() {
+   public int[] getUserIds() {
       return userIds;
    }
 
-   public void setUserIds(long[] userIds) {
+   public void setUserIds(int[] userIds) {
       this.userIds = userIds;
    }
 }
